Made location values int

+++ OnsetJava-API/src/main/java/net/onfirenetwork/onsetjava/api/util/Location.java
@@ -8,20 +8,20 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @Getter
 @Setter
-public class Location extends Vector3d {
+public class Location extends Vector3i {
     double heading;
 
-    public Location(double x, double y, double z, double heading) {
+    public Location(int x, int y, int z, double heading) {
         super(x, y, z);
         this.heading = heading;
     }
 
-    public Location(double x, double y, double z) {
+    public Location(int x, int y, int z) {
         super(x, y, z);
         this.heading = 0;
     }
 
-    public Location(Vector3d vector, double heading) {
+    public Location(Vector3i vector, double heading) {
         super(vector.getX(), vector.getY(), vector.getZ());
         this.heading = heading;
     }
